+++ src/Pages/Topsidenavbar/Side-N-Topbar.js
@@ -36,8 +36,28 @@ const adminsidebar = ({userData}) => {
 
 
 const logOut = () => {
-  window.localStorage.clear();
-  window.location.href = "./sign-in";
+  fetch("http://localhost:5000/logout", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+      token: localStorage.getItem("token"), // Send the token in the "token" header
+    },
+  })
+    .then((response) => response.json())
+    .then((data) => {
+      if (data.status === "ok") {
+        // Clear local storage and redirect to sign-in page
+        console.log("Logout Succesfully");
+        window.localStorage.clear();
+        window.location.href = "./sign-in";
+      } else {
+        // Handle any error response
+        console.log("Logout error:", data.error);
+      }
+    })
+    .catch((error) => {
+      console.log("Logout error:", error);
+    });
 };
 
 const dashboard = () => { 
@@ -66,7 +86,7 @@ function CreateExam() {
 }
 
 function ViewStudentResults() {
-  navigate("/dashboard/viewAllStudentResults");
+  navigate("/dashboard/viewResultsTestsLists");
 }
 
 function CreateUser(){
